test: integration tests for complex endpoints
