+++ src/lib.rs
@@ -6,4 +6,3 @@ pub mod routes;
 //     pub use crate::domain::dates::*;
 //     pub use crate::routes::*;
 // }
-
